Oops forgot this file.....
First phase of cleanup of the dumper(hdp). Significantly Enhanced the error
catching/reporting/cleanup of dumping of low-level objects, SDS, Vdatas
and Vgroups. All that remains are images(DFRxx,GR) for next time. Still
needsanother pass or two to update the comments and verify certain sections
of the code.  -GV 9/24/97