Author: pharten
Problem:  The elements of an attribute array were loaded to incorrect spots
in memory during SDstart().  Because of this, when dumped, the second element
of the attribute array which should have been -1.0 was 0.0.

Solution:  Make sure when loading attribute elements into memory, it depends
on machine size of the variable being loaded.  This worked out to be that
DFKNTsize(HDFtype) was replaced with DFKNTsize(HDFtype | DFNT_NATIVE).