Author: pharten
Problem:  The attribute sizes returned by VSattrinfo() and Vattrinfo() are
the local machine sizes of these fields.  Since machine sizes vary among some
machines, the dumps also vary.

Solution:  A way around this problem is to not use the size returned by
VSattrinfo() and Vattrinfo(), but to use the size of the fields when they are
written out to file.  This is given by the new functions VSattrhdfsize() and
Vattrhdfsize().