Author: pharten
Problem:  The field sizes returned by VSinquire() are the local machine size
of these fields.  Since machine sizes vary among some machine, the dumps also
vary.

Solution:  A way around this problem is to not use the size returned by
VSinquire(), but to use the size of the fields when they are written out to
file.  This is given by the new function VShdfsize().