Problem: While fixing bug 171, I noticed that GRluttoref is not in the test program. I'd like to confirm my change anyway.
Solution:
Added a test for GRluttoref to the GR test program in the file mgr.c
The test includes two calls to GRluttoref, one returns a valid reference number and
the other returns a 0 due to an invalid palette id being passed in the routine.

Tested on Solaris and Solaris X86.