Some SGI machines run IRIX64 but have CPU's that does not support MIPS4 instruction sets.  Current configuration hardwired it to -mips4.
Solution:
Removed the '-mips4' option so that code provided is no longer
hardwired to the MIPS4 instruction set.  It uses whatever the
local CC default or the users' own specification is.  This gives
an extra benefit that for whatever reason (maybe the reason that
MIPS3 object code is smaller than MIPS4), a user may choose to
use MIPS3 code even if the CPU can support MIPS4.