Purpose: Fix mac external element test
Problem: The directory access on the mac was not handled well.

Solution: In opendir(), refnum & directory ID's passed to hopendir if applicable. In hopendir(), some cleanup was added to deal with ':'s. And after a successful PBHSetVol() a PBHGetVol() is done to get the proper new directory ID. There has also been some changes to eliminate spurious warnings.

Platform Tested: Mac & Baldric.