4r1.2 VMS Port
This command file is added for 4r1.2 to build and test the HDF Library on
Alpha VMS and VAX VMS platforms.

It does the following:

creates the directories [.bin], [.lib], and [.include] at the top level
($HDF_SOURCE)

builds jpeg library, copies *.h files and the jpeg library to the [.include]
and [.lib] directories respectively.

builds the zlib, and copies the files

builds full df library and runs testhdf and fortest; copies the files to the
[.include] and [.lib] directory

builds utilities

builds and tests the xdr library

builds and tests the mfhdf library; copies files to the [.include] and [.lib]
directories

builds and tests ncdump, ncgen, and hdp utilites and copies to the [.bin]
directory

necessary cleanup is done at each step