Added projects to build gif2hdf and hdf2gif utilities.
Tested on WindowsNT (janos) and Windows98 (moskva).
mfgrpf.c mfgrpff.f -  added support for JPEG compression to the grscompress
                      Fortran function. Tested on WinNT and 98 platforms.

hfilepf.c hfilepff.f - On 04/17/2000 I added heprnt and hestring Fortran functions.
                       This change broke the sorce code on both WindowsNT and 98
                       platforms. I am reversing the change to start al over
                       again. Tested on WinNt and 98 platforms.

I made some changes to the test_grcompress subroutine to make sorce code on the
Windows platforms to compile. Type of the data was changed from
integer*2 to integer. When datatype is integer*2, linker cannot find mgwrimg
function in the library. I think it is a bug in the test code. Equivalence statement
should be used and integer buffer should be always supplied to the mgwrimg function
on Windows platfoms. Since this program tests grscompress call, I decided that it is OK
for now. I am filing bug report to add test for grscompress with different datatypes.

Tested on WinNT/98 and Solaris 2.6.