Purpose: "bug" fix
Description:
the EOS library writes SDS datasetss that are detected by the HDF4 library as "lone" dimension scales SDSs.
this is due to a call to SDsetdimname at index 0 with  a name that is the same as the  name of the associated SDS.
the side effect of this (a HDF4 bug) is that the SDS is marked as a dimension scale.

Thus, hrepack skips this SDS in its copy routine (it skips using the SDiscoordvar function).
This patch uses the following algorythm to reproduce the "lone" dimension scale:

1) build a list of all SDSs dimension variables using SDfileinfo cycle
2) build a list of all dimension variables inside the cycle for vardims of an SDS
3) compare the 2 lists using a match algorythm
4) if any of the items in 1) is missing in 2) save those SDSs as  dimension scales doing the call to SDsetdimname
with the same name as the SDS


Solution:

Platforms tested:
linux

Misc. update: