Purpose: Redstorm (Cray XT3) port.
Description:
The buffer test could take up to an estimated 9 hours to run in NFS
space.  If done in Lustre space, estimate is 1 hour.  This is way too
long and has an appearance of infinite looping since no output appears
during the long running.

Detailed Changes:
Changed the ELEMSIZE from 16348 to 1000 if it is for the CRAY XT3 so
that the test could finish in about 3 minutes if run in Lustre space.
Also print the ELEMSIZE in the verbose message.

Platforms tested:
Redstorm.