Purpose: Bug fix (652)
Description:
hdf/test/testhdf has a sub-test called buffer. It has several issues:

1. It actually is a performance measurement program.  It uses clock() 
which returns processor (CPU) time. Since it is measuring I/O performance, wall
clock time should be used.  This needs to be fixed.

2. It runs a long time.  The unbuffered read/write takes 99.5% of the time,
as measured by clock().  It should be changed to be a separated test, with
command line option to control the size of tests.  By default, it should
run with a small size just to verify the program works.

3. It should always report the measurement result.  By moving it out of
testhdf to be an independent program, it would permit it to print its own
output.

Solution:
1. Replaced clock() by timeofDay().
2. not covered yet.
3. Reports measurement result.

Tested:
in osage.