bug fix for generation of "lone" dim scales
2 lists of dimension scales are generated. One during a check of
SDiscoordvar for the SDS (if returns true), another during the
generation of the dimensions for each SDS. 
Then the 2 lists are matched to see which names of the first list are
not on the second list. these names are the "lone" dimensions, and they
created as SDSs

One caveat is that during the insertion of the match table, a check must
be made to see if the name is already on the table. If yes, then
information is added at that position instead of adding a duplicate
name.

Moved all the code related to the regeneration of the "lone" dimensions
to hrepack_dim.c and defined a custom table for dimensions (a name and a
ref) and a custom made table for the match of the previous dimension
table.