Fix for buzilla 712 and 713
Regarding #712. There was a logic error on the SDS initialization regarding chunking parameters (the GR was ok tought). The default values for chunking were not set to the ones read from the object.

Regarding #713. The logic of the print information also had an error. Defined some temporary variables regarding if we are on the first or second traversal of the file