Problem:
Compilation of Fortran libraries failed on Windows.

Solution:

Added section for Windows that uses old style FNAME definitions for the C stubs functions.
FNAME macro is not used on the Unix side anymore (F77_FUNC is used instead). 
We will need to clean up header files after we make sure that everything works.

Platfroms tested: 

kagiso and Kent's Winodws machine.

Note:
We are in the process of making HDF4 to work on Windows; this approach may change
if we find a better solution.