Purpose: Upgrade.
Description:
The daily test failure of copper has been traced to the use of "rsh copper .."
in launching the bin/snapshot.  When "ssh copper ..." is used instead, it
runs fine.

Solution:
Changed to first use "ssh" for connection attempt.  This is similar to
the change in hdf5 runtest.

Platform Tested:
in osage to call copper.