Purpose: Update cygwin documentation for Fortran support
Description:
We have found a "solution" to a bug in cygwin configure that allows us to support building Fortran libraries.  The fix involves overriding the make variable 'EXEEXT' with an environment variable.  This tricks make into using g95 correctly.

Although this is a hack, it will be better than having no Fortran support in cygwin.

Tested:
Cygwin on WinXP