Purpose: Depreciate Win32.zip
Description:
We are beginning a project to update the Windows portion of HDF4 to make it easier to maintain.  The first step is to get rid of Win32.zip.  This checkin creates a new top-level folder, windows.  Just like HDF5, we now have a copy_hdf.bat file which copies all relevant code to it's appropriate place.

I've also updated the documentation and MANIFEST to reflect these changes.  However, we don't have chkmanifest on Windows, so there's a good chance I missed something.  I'll correct it in another checkin.

Tested:
VS6 on WinXP