Purpose: Update misc. paths from yesterday's change
Description:
In depreciating Win32.zip, the addition directories added required that I move a few of the project files (into the ./windows/proj folder).  Because some of the project settings use relative paths, this changed the output directory.  As a result, we need to update files that use this output: a few other projects, and our install scripts.

Tested:
VS.NET on WinXP