hdiff detection of "lone" dimension scales new test at shell script, 2 new files checked previously generated with a previous HDF4 version , that have a lone dimension
The same algorithm used by hrepack was implemented for hdiff for the detection of lone dimension scales. In this case however, four tables are needed instead of two.

Table 1. File 1 filled with SDiscoordvar
Table 2. File 1 filled with SDS dimension cycle with SDdiminfo
Table 3. File 2 filled with SDiscoordvar
Table 4. File 3 filled with SDS dimension cycle with SDdiminfo

The same match is done for each file: find the lone dimensions. Then a further match is done to see if the lone dimension also exists in the second file. Then the difference between 2 SDSs is made as "usual"

tested: linux, sunos 5.10, windows