Purpose: Restructure and cleanup Windows files
Description:
This is a major restructuring of the Windows projects and output.  The hope is to cleanup our project files to minimize duplication, and to generate output in a way that is intuitive and mirrors other platforms.  Previously, Windows HDF4 has been very "messy"-- this addresses much of that mess, and streamlines the build process as well.

Specifically, this checkin makes the following changes:
-- Puts Windows project files in a directory structure rooted at ./windows/proj/
-- Changes project output and directory structures to correspond with source directories
-- Update batch scripts accordingly
-- Misc. project cleanup (and add Pedro's new source file to hrepacktest)

Notes:
-- I expect this will probably break the MANIFEST.  I've done my best to update it, but we don't have chkmanifest on Windows, so I've probably missed something.
-- I also haven't updated the documentation.  We are going to split VS6 on V.NET later this week, so many things are still going to change.  At that point, there will be a major re-write to the Windows Install file.

Tested:
VS6 on WinXP
VS.NET w/ Fortran on WinXP