Purpose: Rearrange Windows example project files
Description:
We have decided to maintain example projects and source on Windows for now (we can move them to other platforms when we work on restructuring the rest of the source).  This update puts all of the example project files in one folder, "proj", which will make it easier when we split vs6 and .NET.

Also cleaned up empty ./windows/fortran folder that was left behind from last checkin.

Tested:
VS6 on WinXP