Purpose: Split Windows project files for Visual Studio .NET
Description:
We would like to eventually depreciate the Visual Studio 6.0 compiler and move our project files to more recent compilers.  As part of that process, we have split our projects into files for VS6.0, and VS.NET and above.  This is neccessary for Fortran support, and allows us to streamline the build process for users of VS .NET or 2005 (which is most).  Specifically, this checkin makes the following changes:

-- Split our library and example projects into two folders-- proj and proj_vs6.
-- Split the library solution into two-- all.sln and all_fortran.sln, which includes converted Intel Fortran project files.
-- Many misc. fixes and cleanup for Windows.

Notes:
-- This change neccessitates almost a complete re-write of our Windows documentation.  This means that the current documentation is out of date.  I will update it within the week, but if you need help in the meantime, see me.
-- Windows does not have chkmanifest so, while I've tried to update the MANIFEST accordingly, I've probably missed something.  I will check it and fix it accordingly.

Tested:
VS6
VS.NET
VS2005