Purpose: Major Windows documentation, and misc. enhancements/cleanup
Description:
This brings the Windows documentation fully up-to-date, including new instructions for 6.0 and .NET, revised instructions for 2005, and new Fortran support.  I've also been able to streamline some of the testing in .NET projects (ctest0).  And the project output for C and Fortran examples have be moved, with batch script updates accordingly.

Tested:
VS.NET
VS6.0