Purpose: Bring Windows command-line build up to date.
Description:
This is a re-write of the HDF4 Windows commandline build script, similar to what's done on HDF5.  Using this script, users can now build from the commandline with the VS.NET or 2005 compilers, and with or without Fortran.  There is now also much more error checking.  

I've updated the documentation, and also made minor edits to other scripts to make them a little more uniform.

Tested:
VS.NET
VS2005