Purpose: Update hdp test script with recent additions
Description:
New tests were added to the testhdp.sh.in-- this adds those, as well as others that were mysteriously missing.

Note that on Windows, we comment out testing of debug versions, as they seem to fail.  This is still the case.  Specifically, tests 12, 18, and 19 from the "dumpgr" series.

Tested:
VS.NET on WinXP