Purpose: Remove Windows auto-linking
Description:
In HDF4, we previously used preprocessor logic to determine the build configuration, and link accordingly.  This was bad, because it uses hard-coded values, and hides the "correct" way to do it-- using project dependencies and explicitly setting link dependencies.

I've removed this Windows conditional logic from hdfi.h and mfhdf.h, and updated example project files that were depending on it.  Everything else works as usual, and will be "safer" in the future.

Tested:
All Windows projects, VS6 and VS.NET
Not tested on Linux, because all changes were wrapped in a Windows-specific "#ifdef"