Purpose: Fix Windows HDF-SD C interface tests
Description:
Recently, hdftest was updated to try and find the testfiles in a new way that doesn't work with Windows.  Windows doesn't set $srcdir as other platforms do, and it also uses different relative paths.  We've added a special-case to make sure that we can find our input.

Tested:
VS6 on WinXP
Wrapped in "#ifdef _WIN32"; won't affect other platforms.