Purpose: Fix subtle "FC" bug in Windows test scripts.
Description:
On Windows, the "FC" command (which is like "diff" on unix) doesn't return an error code to indicate pass/fail.  As a result, we pipe the results to the "find" command, searching for the string "FC: no diff", which is output from FC when the files are identical.  Previously we checked the error-level after calling FC, which has unpredictable results.

Tested:
VS.NET on WinXP