Purpose: Modify WIN32 ifdefs for compatibility with 64-bit Windows
Description:
On 64-bit Windows, the preprocessor _WIN32 is automatically defined for backwards compatibility.  However, the macro WIN32 is not necessarily.  In most of our project files, we manually make this definition, but we shouldn't depend on it.  This changes all references from WIN32 to _WIN32.  

This shouldn't break anything because other platforms still won't define it, and all supported versions of Visual Studio will.

Tested:
VS2005 on WinXP-32