The previous fix will cause special build for GE on windows 64-bit failed. Since GE would like to use DLL flag for static library and programs, The previous _DLL will be passed to the corresponding projects and build DLL instead of static programs.
If we just comment out define _HDFDLL_ in hdfi.h, this will make 64-bit GE windows build pass; however, it will cause 32-bit fortran release DLL fail since somehow although we pass _HDFDLL_ to be a predefined preprocessor, MSVS2005 doesn't acknowledge it. It still tried _HDFDLL_ as undefined. It won't go to the following piece of code and generate fortran stub DLLs with the exporting library. It seems a compiler bug. But we have to face it. So we will use a special header file for GE build for the time being and will find a good solution later.

#       if defined(_HDFDLL_)
#               pragma warning( disable: 4273 ) /* Disable the stupid dll linkage warnings */

#               if !defined(_HDFLIB_)
#                       define HDFPUBLIC __declspec(dllimport)
#               else
#                       define HDFPUBLIC __declspec(dllexport)
#               endif

#               if !defined(_MFHDFLIB_) && !defined(_HDFLIB_)
#                       define HDFLIBAPI __declspec(dllimport) extern
#               else
#                       define HDFLIBAPI __declspec(dllexport) extern
#               endif

#               if defined(_HDFLIB_C_STUB_EXPORTS) || defined(_MFHDFLIB_C_STUB_EXPORTS) || defined(_DLLLIBTEST_FCSTUB_EXPORTS)
#                       define HDFFCLIBAPI __declspec(dllexport) extern
#               else
#                       define HDFFCLIBAPI __declspec(dllimport) extern
#               endif