Purpose: Add Fortran szip support for Windows
Description:
Windows uses its own modified versions of Fortran source files in the fort_ps folder.  When Fortran szip support was added to the main source, it was left out of the Windows version.  This checkin brings the Fortran szip support to the Windows.  This is essentially a merge of the changes from rev. 4703 and 4712 to Windows.

Tested:
VS2005 on WinXP (tszip test only)
No other platforms tested, because only Windows source code affected