Purpose: In Windows command line build script, detect installed Visual Studio version with Intel Fortran
Description:
Although we support both Visual Studio .NET and 2005 with Intel Fortran 9.1, Intel Fortran must be targeted to one or the other when installed, and only the targeted version of Visual Studio can be used with Fortran on the command line.

We previously tried to detect this before building, but the script had a few errors. This checkin fixes those errors and makes the detection process much cleaner.

Tested:
VS.NET on WinXP