Purpose: Add new definition H4_HAVE_SZIP_ENCODER to Windows
Description:
A new macro, H4_HAVE_SZIP_ENCODER was added to the general h4config.h.in and configure scripts.  We don't use these on Windows, so we need to add the definition to the maintained h4config.h file.  The macro is defined to '1' by default.

When a user wants to use szip without encoder, they will need to manually update the h4config.h file to specify this change.  This has also been described in the INSTALL_Windows.txt document.