Description:
Fixed scgetfname to allocate a C character buffer in order to hold the returned file name and allocate it size +1 to account for Null termination. Fixes problem of crashing on kagiso and ifort when piping output to a file.

Fixed scname2ind: made n_vars *n_vars and changed array index.