Purpose: Update new test, hdfnctest, on Windows
Description:
There was a new test added recently, hdfnctest.  This checkin fixes a few minor issues with the Windows version, and adds the test the complete test batch file.

Note that the test has not yet been added for Visual Studio 6 or Visual Studio 2003-- I'll work on adding these configurations on Thursday.

Tested:
VS2005 on WinXP