PurPurpose: Add new project files for test hdfnctest for other Windows compilers
Description:
There was a new test added recently, hdfnctest.  Previously we only created the project for Visual Studio 2005.  This checkin adds the necessary files for Visual Studio .NET and 6.0.

Note: Currently the tests for Visual Studio .NET fail, but I have emailed Binh-Minh to see if we can work it out.

Tested:
VS.NET on WinXP
VS6 on WinXP