Purpose: Update the version string for Windows generated DLLs.
Description:
On Windows, the HDF4 library files we generate contain a version string in the filename of the output file.  The version string needs to be updated to reflect the recent release.  The string is only present in 4 output libraries (hdf + mfhdf, static + release).  However, many other example projects use the library as linker input, and also needed to be updated.  Finally, we update the installation batch scripts to use the new name as well.

Tested:
VS2005 on WinXP, build + install only.