Purpose: Remove a 'hack' which was required for building Fortran under Cygwin
Description:
About a year ago, a Cygwin update broke Fortran support for HDF4 because of the way certain Makefile syntax was interpretted.  Specifically, the EXEEXT wasn't being properly recognized on the fortestF project, and as a result the incorrect compiler and options would be used.  We came up with a work-around, but it is non-standard and very confusing for users.

In this checkin I've found a way to coerce Cygwin to correctly interpret the Makefile without the hack.  Basically, we hard-code the $(EXEEXT) variable into the Makefile.am dependency.  Now users can build normally without worrying about special Fortran instructions.

This checkin changes the Makefile.am, and I've also re-run bin/reconfigure to propogate the changes to Makefile.in.  Finally, I've removed the additional 'hack' from the INSTALL_CYGWIN.txt instructions.

Tested:
cygwin, jam