Purpose: Minor updates to Cygwin install documentation
Description:
Update the Cygwin documentation based on some problems we found when trying to build in a minimal install environment.  Updates include:
- List more explicitly which Cygwin packages will need to be installed beyond that of a minimal Cygwin installation
- List the download URL for the g95 compiler
- List the download URL for the szip library
- Note that we currently cannot build HDF4 without any Fortran compiler installed

Also removed the exact version string from a couple places in the documentation where it is unnecessary.  Just so we have one less place to update it during releases.