Description:
The tests now only use copies of the *.dat files (distributed in the source) in the build directory. That way, the files in the source directory will not be modified when building the code in another directory. It also fixes the problem when the source directory is set for read-only, which would cause the tests that try to modify the *.dat files to fail. It copies the *.dat files for "make check", and changes the permission of the *.dat files to a+rw. It then removes the files for make distclean.

Tested smirom, honest.