Descriptioni:
The bug mentioned in the comments for tanf.f:

C  Possible bug:  When reading in a label, we have to give it a 
C                 length that is one greater than MAXLEN_LAB. This
C                 may be due to a bug in dfan.c in DFANIgetann().

and noted in bug 1640 was fixed by adding a isfortran 
flag to DFANIgetann to avoid termination 
of the string with the null character when the routine 
is call from FORTRAN, hence no longer need to add one.

Tested: jam (ifort/icc), gcc/g95