Bug fix: HDFFR-1318, HDFFR-1327 and HDFFR-1328
Descripton:
Encountered seg fault and test failures in Mac OS X Mountain Lion, same as
in Lion.

Solution:
Dropping the C compiler to no optimization (-O0) eliminated the errors
by setting $PROD_CFLAGS to -O0 if it is for a gcc and in a Mountain Lion
(darwin12.*) or a Lion (darwin11.*).  This will in turn set CGLAFS to -O0.
Moved the fix into Apple and eliminated the more specific apple-darwin11.4.*
files. This should work for future different versions of Darwin11.X and
Darwin12.X.

Tested:
Duck, Fred, Owl (Darwin12). And passed H4committest.