Maintenance: This check-in is to minimize an affect of HDFFR-1477 on the output of the Fortran examples.
For GR and VG/VD interfaces we do not have functions to return the length of the strings,
such as attributes names or names of the GR images. As a result, Fortran application has to 
pass a buffer big enough to get the whole string. If the Fortran buffer was not initialized,
C puts some garbage in it after the real string values. The garbage is displayed as a string
of "^@" characters as the examples produced. The examples were modified to pass initialized
strings. Only end of the C string character is printed now, making output legible.

Bug will be addressed in the examples as soon as we have necessary functions that return
the length of a string stored in HDF4.

Platforms tested: jam, platypus with gcc, intel and pgi compilers, emu with sun compilers.