Fixed HDFFR-1593 (#159)
* Fixed HDFFR-1593

The number of fields in a vdata from the user's file was corrupted
and exceeded the maximum allowed.  Continuing the process with it
will likely cause segfault at a later time.  This commit added checks
for both dumpvd and dumpvg to fail when such number of fields was
read from the file.

* Update hdp_vd.c

* Update hdp_vg.c

* Update hdp_vd.c

Co-authored-by: Dana Robinson <43805+derobins@users.noreply.github.com>