Quiets 'ar' warnings from -u option (#252)
The -u option (default in Automake) only updates changed files in the archive.
Leaving it out makes packing a bit slower, but ensures a fresh archive every
time. The default in GNU ar is to use -D ('deterministic') which creates
a fresh archive each time.

It's better for us to use the GNU default of -D than to use the Automake
default of -u. Droppig the -u also quiets a complaint from ar that -u
is incompatible with -D.