Fix "failing" netCDF tests (#396)
A couple of tests in tncvarget would emit messages about reading bad
values, but would not actually fail the test. This change fixes
all the -Wshadowed warnings, fixes the array code so the tests
pass, and properly check the number of errors so the tests will
fail on errors.

Note that the library did not actually do the wrong thing - these
were just poorly-coded tests.