Final 4 2 16 patch changes (#399)
* Fixed HDFFR-1520 (#334)

Applied user's suggestion with further improvement to conform to the
library's way of validating the file record.

* Add support for kFreeBSD & Hurd architectures. (#337)

* Remove the -ansi flag from clang compile options (#341)

This sets C89 mode and overrides -std=c99, generating a ton of warnings.

* Fix failures in xdr test (#342)

* Fix failures in xdr test

- Corrected the parameter list of xdrproc_t
- Casted the function pointer in calls to xdr_vector to eliminate
    "incompatible function pointer types" warnings.

* Added detailed comment for xdrproc_t

* Fix formatting

* ci: add HDF-EOS2 test action (#343) (#344)

* Fix H4_DFI_H endif. (#366)

* Fully remove partially-deleted APOLLO #ifdef code
* Remove 'Machine dependencies' block

* Update macOS GitHub actions to 13 (#368)

* Update the gcc version to 12 in GitHub actions (#375)

* Fix gfortran version check (#377)

The config/gnu-flags code checked for the gfortran version using
a variable that was never set, so -fallow-argument-mismatch was
never set for gcc 10+, which prevents building the Fortran
libraries.

Fixes HDFFR-1582

* Move all 'warnings as errors' to warnings (#379)

This scheme was brought over from HDF5 and only worked in CMake.
It breaks Fortran, though, since some of those files are *really*
old, so all the 'warnings as errors' are now just regular warnings.

Also brings the warnings scheme into the Autotools (gcc only)

* Remove duplicated CFLAGS (#380)

* Add Fortran to CI (#384)

CMake only (there's a linking issue w/ Autotools)

* Fix CFLAGS when using the system RPC (#393)

The TI-RPC, etc. checks unset the CFLAGS variable in the feature
tests.

* Fixed HDFFR-1609 (#369)

Fixing HDFFR-1607 introduced a bug where the flag that indicates whether
the directory of the external file has been changed from the previous
location was not reset properly.

Reset the flag after building a filename path for an external element.

Fixes HDFFR-1609

* Fix "failing" netCDF tests (#396)

A couple of tests in tncvarget would emit messages about reading bad
values, but would not actually fail the test. This change fixes
all the -Wshadowed warnings, fixes the array code so the tests
pass, and properly check the number of errors so the tests will
fail on errors.

Note that the library did not actually do the wrong thing - these
were just poorly-coded tests.

* Test code improvement in tncvargetfill (#397)

- Added a verification of number of records in NC API test
- Fixed declaration/initialization of a couple arrays
- Cleaned up comments in test
- Improved memory allocation

* Regenerated Autotools files

* Fix codespell issue

* Update GitHub action branch