Incorporated XDR into the mfhdf library
We no longer build a separate XDR library as all the XDR functionality
has been moved to mfhdf/libsrc and is built with the mfhdf library.
Because of this, we no longer distribute a separate XDR library with
either build system.

We also no longer distribute the XDR types.h and xdr.h headers.