Fixed SZIP availability check  (#510)
* Fix HDFFR-1529

The line hcomp.c:157 fixes this issue.  The rest of the changes fix typos.

* Fixed dumper for no SZIP case

When the SZIP library is not available, HDF4 library is unable to retrieve the SZIP c
ompression information.  This issue was revealed after the fix of HDFFR-1529 because before that, the HDF4 library did not detect that SZIP library was unavailable.

hdp now displays appropriate information when SZIP library is not available and
stops trying to obtain SZIP compression information.