Fix encoder detection w/ --with-szlib & no paths
The Autotools --with-szlib option did not correctly detect the
presence of the encoder due to a mis-named variable. When enabled,
a missing variable would cause the dumper tests to fail due to
picking the non-szip file for output diffs.

This is fixed now, and --with-szlib works as expected.

Also cleans up some szip-related warnings.