Fix an example
An incorrect input file was used for this example.  The example
h4ex_VD_write_to_vdata.c generated the file General_Vdatas3.hdf
but this example used General_Vdatas.hdf, which didn't have the
data the example expected, which was the same as in h4ex_VD_write_to_vdata.c

Also changed to return an error when the vdata_ref was not found.
However, unless the correct file is not available or somehow
corrupted, this should never happen.

Addressed HDFFR-1543