Fix memory leaks in Vgroup and Vdata APIs
The code that destroys the atom groups for Vdatas and Vgroups was placed
after the free lists had already been freed, so, that code added more
released memory to the free lists, which were never freed, causing many
memory leaks.
This PR moved that block of code appropriately and removed all of those leaks.

In addition, this PR:
- added checks after mallocs
- fixed typo HAremove_atom, should be HAatom_object
- removed vgroup from atom list after it is detached

Fixes GH-819